added animations and sound effects
Idle and Turn animations added for the remaining 2 characters.
Added temporary sound effects which play upon taking damage, receiving healing, death, missing with an ability and when there's an error.
Changed the way characters are spawned - the BattleHandler takes the positions as Transforms and keeps them in arrays.
New background image just because.
Started working on a text popup which will be used to indicate characters communicating.