fixed some buffs and added buff bar
Buffs now come in three variants, depending on when they "tick" - at the start of the character's turn (timed buffs), whenever the characters makes an attack (on attack buffs) and whenever the character takes damage (on take damage buffs).
Buffs now separated in 3 different lists depending on variant.
Buffs have enum type, which is used to display their icon on the buff bar (icons are stored in the game assets object).
Added tooltips for each buff.
Character now has bool isTargetable, which is changed to "false" when a character dies or when it is "hidden" by a buff.
Fixed some issues when trying to access null pointers.
Fixed enemy target picking to not endlessly loop when all remaining characters are untargetable.